backend/requirements.txt
asgiref==3.7.2
Django==3.2.19
django-filter==23.2
djangorestframework==3.14.0
djoser==2.1.0
flake8==5.0.4
Pillow==9.5.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
zipp==3.15.0
